% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspc}
\alias{gspc}
\title{Daily log return of the GSPC from 2004-01-05 to 2024-08-02}
\format{
A vector with length 5180.
\describe{
  \item{log return}{log return}
  ...
}
}
\source{
\url{https://finance.yahoo.com/quote/\%5EGSPC/history?period1=949363200&period2=1025222400&interval=1d&filter=history&frequency=1d}
}
\usage{
data(gspc)
}
\description{
This is processed from the original ^GSPC.csv dataframe.
}
\details{
GSPC is a ticker for S&P 500 Index, see \url{https://en.wikipedia.org/wiki/S\%26P_500#cite_note-10}.
}
\keyword{datasets}
