% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ret_gncma.R
\docType{data}
\name{retGNCMA}
\alias{retGNCMA}
\title{Daily Return of General Communication from 2006-01-06 to 2010-12-31.}
\format{
A xts object with length 261 weekly data for 1 variable.
\describe{
  \item{log return}{log return}
  ...
}
}
\usage{
data(retGNCMA)
}
\description{
Daily Return of General Communication from 2006-01-06 to 2010-12-31.
}
\keyword{datasets}
