% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-fit.R
\name{rg}
\alias{rg}
\alias{fit}
\alias{nnest1}
\alias{nnest2}
\alias{varin}
\alias{tausq}
\alias{sest}
\alias{rho}
\alias{gk}
\alias{freg}
\alias{Fnue}
\alias{Fnue2}
\alias{nfun}
\alias{nfun2}
\title{Robust GARCH(1,1) Model Estimation}
\usage{
fit(x, bms = 1)
}
\description{

}
