% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-mest.R
\name{rgMEst}
\alias{rgMEst}
\alias{rgmest}
\title{Robust Estimates for GARCH(1,1) Model}
\usage{
rgMEst(
  data,
  classes = c("bounded MEst", "modified MEst"),
  class_pars = c(0.85, 3),
  optimizer = c("fminsearch", "nloptr", "Rsolnp"),
  optimizer_control = list(),
  stdErr_method = c("finite difference hessian estimator",
    "outer product of gradients", "sandwich estimator")
)
}
\arguments{
\item{data}{a time series of log returns}

\item{class_pars}{a named numeric vector of parameters of the class. For the "modified MEst" class, the parameter should be "div", which is used in m(rho0(x)) = div * m(rho0(x)/div) and defined as 0 <= div <= 1.0, default is div = 0.85; for the "bounded MEst" class, the parameter should be "div" and "k", which also influences the robustness and defined as k >= 0, default is div = 0.85, k = 3. The smaller div or k is, the more robust the moddel is.}

\item{optimizer}{optimizer used for optimization, one of "fminsearch", "nloptr", "Rsolnp", default is "fminsearch"}

\item{optimizer_control}{list of control arguments passed to the optimizer}

\item{stdErr_method}{method used to calculate standard error, one of "finite difference hessian estimator", "outer product of gradients", "sandwich estimator", default is "finite difference hessian estimator" using hessian from numDeriv}

\item{class}{robust M-Estimate class used for Garch(1,1) model, one of "modified MEst", "bounded MEst", default is "bounded MEst"}
}
\value{
A \code{rg} object(S3), the components of the object are:
    \item{data}{the log returns data object for the rg model to be fitted}
    \item{classes}{the class called}
    \item{class_pars}{named numeric vector of parameters for the class}
    \item{optimizer}{the optimizer called}
    \item{optimizer_control}{the list of control arguments passed to the optimizer}
    \item{optimizer_result}{output of the called optimizer}
    \item{stdErr_method}{the method called to calculate standard error}
    \item{fitted_pars}{Garch(1,1) parameter estimations output of the called class}
    \item{objective}{the optimal likihood value obtained by the optimizer}
    \item{time_elapsed}{the time used for the optimization routine}
    \item{message}{the message of the convergence status produced by the called solver}
    \item{standard_error}{standard erros of the fitted parameters using the method called}
    \item{t_value}{t-values of alpha_0, alpha_1, beta_1}
    \item{p_value}{p-values of alpha_0, alpha_1, beta_1}
}
\description{
Methods for fitting a Garch(1,1) model with daily log return time series, using two classes of robust extended M-Estimates:
(1) maximizing modified likelihood function;
(2) M-Estimates with bounds to the propagation of the effect of one outlier on the subsequent predictors of the conditional variances.
returns a \code{rg} object
}
\details{
The \code{rgMEst} function fits a Garch(1, 1) model to a time series of log return data, using one of the two classes of robust extended M-Estimates with certain parameters specified by the user, with guidance and examples from the vignette. The user can also specify the optimizer used during optimization procesure, and the method used to calculate standard error for the fitted parameters.

For details of the list of control arguments, please refer to \code{fminsearch}, \code{nloptr::nloptr}, \code{Rsolnp::solnp}
}
\examples{
# TBD.








}
\references{
Muler, Nora & Yohai, Victor. (2008). Robust estimates for GARCH models. Journal of Statistical Planning and Inference. 138. 2918-2940.
}
