% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-package.R
\docType{package}
\name{robGarch-package}
\alias{robGarch}
\alias{robGarch-package}
\title{robGarch: Robust Garch(1,1) Model}
\description{
Estimation and methods of two Robust Garch(1,1) models
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EchoRLiu/robGarch}
  \item Report bugs at \url{https://github.com/EchoRLiu/robGarch/issues}
}

}
\keyword{internal}
