% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustGARCH-package.R
\docType{package}
\name{robustGARCH}
\alias{robustGarch}
\alias{robustGarch-package}
\alias{robustGARCH}
\title{Robust GARCH Package}
\description{
Robust GARCH modeling functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EchoRLiu/robustGarch}
  \item Report bugs at \url{https://github.com/EchoRLiu/robustGarch/issues}
}

}
\author{
\strong{Maintainer}: Echo Liu \email{yuhong.echo.liu@gmail.com}

Authors:
\itemize{
  \item Daniel Xia \email{dxia94@gmail.com}
  \item R. Douglas Martin \email{dougPCRA@gmail.com}
}

}
