% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtn}
\alias{rtn}
\title{Daily log return of the GSPC from 2000-02-01 to 2006-01-31}
\format{
A vector with length 5029.
\describe{
  \item{log return}{log return}
  ...
}
}
\source{
\url{https://finance.yahoo.com/quote/\%5EGSPC/history?period1=949363200&period2=1025222400&interval=1d&filter=history&frequency=1d}
}
\usage{
data(rtn)
}
\description{
This is processed from the original ^GSPC.csv dataframe.
}
\keyword{datasets}
