% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-summary.R
\name{rg-summary}
\alias{rg-summary}
\alias{summary.rg}
\alias{print.rg}
\alias{plot.rg}
\title{Summary for rg class}
\usage{
\method{summary}{rg}(fit)

\method{print}{rg}(fit)

\method{plot}{rg}(fit, main_name = "Conditional Volatility (vs |Returns|)", abs_ = TRUE)
}
\arguments{
\item{fit}{A RG fit object of class \code{\link{rg}}}

\item{main_name}{the title of the plot, default is "Conditional SD (vs returns)"}

\item{abs_}{a logical argument, when TRUE, the plot function will plot abs(returns) with conditional standard deviation instead of returns, default to TRUE.}
}
\description{
Summary for rg S3 class
}
\examples{

data("rtn")
fit <- robGarch(rtn, methods="bounded MEst", fixed_pars = c(0.8, 3.0), optimizer="Rsolnp", stdErr_method = "numDeriv")
summary(fit)
print(fit)
plot(fit)

}
