% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-summary.R
\name{rg-summary}
\alias{rg-summary}
\alias{summary.rg}
\alias{print.rg}
\alias{plot.rg}
\alias{aef}
\title{Summary for rg class}
\usage{
\method{summary}{rg}(fit, digits = 3)

\method{print}{rg}(fit, digits = 3)

\method{plot}{rg}(
  fit,
  digits = 3,
  pctReturn_ = TRUE,
  abs_ = TRUE,
  original_ = FALSE,
  main_name = "Conditional Volatility (vs |pctReturns(\%)|)"
)

aef(fit, v = 5)
}
\arguments{
\item{fit}{A RG fit object of class \code{\link{rg}}}

\item{digits}{the number of digits for print and plot, default is 3.}

\item{pctReturn_}{a logical argument. IF TRUE, the plot function will plot the returns in percentage instead of original. Default is TRUE.}

\item{abs_}{a logical argument, when TRUE, the plot function will plot abs(returns) with conditional standard deviation instead of returns, default to TRUE.}

\item{original_}{a logical argument. If TRUE, the original return will be plotted. Default is FALSE}

\item{main_name}{the title of the plot, default is "Conditional SD (vs returns)"}
}
\description{
Summary for rg S3 class
}
\examples{

data("rtn")
fit <- robGarch(rtn, methods="BM", fixed_pars = c(0.8, 3.0), optimizer="Rsolnp", stdErr_method = "numDeriv")
summary(fit)
print(fit)
plot(fit)
aef(fit)

}
