% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg-summary.R
\name{rg-summary}
\alias{rg-summary}
\alias{summary.rg}
\alias{print.rg}
\alias{plot.rg}
\title{Summary for rg class}
\usage{
\method{summary}{rg}(fit)

\method{print}{rg}(fit)

\method{plot}{rg}(fit)
}
\arguments{
\item{fit}{A RG fit object of class \code{\link{rg}}}
}
\description{
Summary for rg S3 class
}
\details{

}
\examples{

data("rtn")
fit <- rgMEst(rtn, methods="bounded MEst", fixed_pars = c(0.8, 3.0), optimizer="Rsolnp", stdErr_method = "numDeriv")
summary(fit)
print(fit)
plot(fit)

}
