% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustGARCH-package.R
\docType{package}
\name{robustGarch-package}
\alias{robustGarch}
\alias{robustGarch-package}
\title{robustGarch: Robust Garch(1,1) Model}
\description{
Estimation and methods of two Robust Garch(1,1) models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EchoRLiu/robustGarch}
  \item Report bugs at \url{https://github.com/EchoRLiu/robustGarch/issues}
}

}
\keyword{internal}
