% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robGarch.R
\name{robGarch}
\alias{robGarch}
\title{Robust GARCH(1,1) Model Estimated}
\usage{
robGarch(
  data,
  methods = c("BM", "M", "QML", "MLE"),
  turingPars = c(0.8, 3),
  optimizer = c("Rsolnp", "nloptr", "nlminb"),
  optimizer_x0 = FALSE,
  optimizer_control = list(trace = 0),
  stdErr_method = c("numDeriv", "optim", "sandwich")
)
}
\arguments{
\item{data}{an xts object}

\item{optimizer}{character valued optimizer name, one of
"Rsolnp", "nloptr", "nlminb", with default "Rsolnp".}

\item{fitMethod}{character valued name of fitting method,
one of "BM", "M" "QML" or "tMLE", with "BM" the default value.}

\item{robtunePars}{a numeric vector c(cM,cFlt) that controls
the extent of fitMethod robustness, with default c(0.8,3.0).}

\item{initialPars}{numeric user-defined initial parameters
c(gamma0, alpha0, beta0) for use by optimizer, with default
values c(0.0005, 0.2, 0.7).}

\item{optControl}{list of arguments passed to optimizer, with
default \code{list(trace=0)}.}

\item{SEmethod}{character valued name of standard error method,
one of "numDeriv", "optim", "sandwich", with default "numDeriv".
 using hessian from numDeriv}
}
\value{
A \code{robustGARCH} object(S3), the components of the object are:
    \item{data}{the input xts object}
    \item{fitMethod}{the the fitMethod specified}
    \item{robtunePars}{the robtunePars specified}
    \item{optimizer}{the optimizer type specified}
    \item{initialPars}{the initialPars specified}
    \item{optControl}{the list of control values specified}
    \item{optOutput}{a list of the optimizer output values returned}
    \item{SEmethod}{the specidied of calculating standard errors}
    \item{coefEstimates}{computed parameter estimates}
    \item{sigma}{conditional standard deviation xts class time series}
    \item{observedInfomat}{observed information matrix}
    \item{objective}{the optimizer minimized objective function value}
    \item{message}{optimizer convergence satus message}
}
\description{
Computes "BM" robust Garch(1,1) model parameter
estimate by using a bounded objective function and a bounded
conditional variance recursion.  Alternatively, it computes:
(1) "M" estimates by using only the bounded objective function,
(2) "QML" estimates based on a typically incorrect assumption
of normally distributed innovations, (3) "t-MLE" estimates based
on an assumption of an innovations t-distributed MLE with unknown
location, scale,and degrees of freedom parameters.
CHECK IF (3) IS CORRECT.
}
\details{
The "BM" fit method delivers the highest robustness by
using a half-Huber psi function to bound the normal distribution
log-likelihood, and using a Huber psi function to prevent the
propagation of influential outliers in the variance recursion.
The "M" method is obtained by dropping the BM bounding of the
variance recursion, and is therefore less robust toward outliers.

Echo or Dan, please provide details for optControl.

The SEmethod default "numDeriv" is based on the Hessian from the
optimization.

The \code{robGarch} function fits a Garch(1, 1) model to a time series of log return data, using one of the two methods of robust extended M-Estimates with certain parameters specified by the user, with guidance and examples from the vignette. The user can also specify the optimizer used during optimization procesure, and the method used to calculate standard error for the fitted parameters.

For details of the list of control arguments, please refer to \code{nloptr::nloptr}, \code{Rsolnp::solnp}, \code{nlminb}.
}
\examples{


data("gspc")
fit <- robGarch(gspc[1:604], methods="BM", tuningPars = c(0.8, 3.0))


}
\references{
Muler, Nora & Yohai, Victor. (2008). Robust estimates for GARCH models. Journal of Statistical Planning and Inference. 138. 2918-2940.
}
