% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustGARCH-summary.R
\name{robustGARCH-summary}
\alias{robustGARCH-summary}
\alias{summary.robustGARCH}
\alias{print.robustGARCH}
\alias{plot.robustGARCH}
\alias{aef}
\title{Summary for robustGARCH class}
\usage{
\method{summary}{robustGARCH}(fit, digits = 3)

\method{print}{robustGARCH}(fit, digits = 3)

\method{plot}{robustGARCH}(
  fit,
  digits = 3,
  estimation_pos = "topleft",
  line_name_pos = "topright",
  par_ = par(no.readonly = TRUE),
  pctReturn_ = TRUE,
  abs_ = TRUE,
  original_ = FALSE,
  main_name = "Conditional Volatility (vs |pctReturns(\%)|)"
)

aef(fit, v = 5)
}
\arguments{
\item{fit}{A robustGARCH fit object of class \code{\link{robustGARCH}}}

\item{digits}{the number of digits for print and plot, default is 3.}

\item{estimation_pos}{string that determines the legend position that specifies gamma, alpha, beta estimations. Choice of "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default is "topleft".}

\item{line_name_pos}{string that determines the legend position that specifies the names of lines in the plot. Choice of "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default is "topright".}

\item{par_}{graphical parameters that can be set, which is in the form of par(...). The default is par(no.readonly = TRUE).}

\item{pctReturn_}{a logical argument. IF TRUE, the plot function will plot the returns in percentage instead of original. Default is TRUE.}

\item{abs_}{a logical argument, when TRUE, the plot function will plot abs(returns) with conditional standard deviation instead of returns, default to TRUE.}

\item{original_}{a logical argument. If TRUE, the original return will be plotted. Default is FALSE}

\item{main_name}{the title of the plot, default is "Conditional SD (vs returns)"}
}
\description{
Summary for robustGARCH S3 class
}
\examples{

data("rtn")
fit <- robGarch(rtn, methods="BM", fixed_pars = c(0.8, 3.0), optimizer="Rsolnp", stdErr_method = "numDeriv")
summary(fit)
print(fit)
plot(fit)
aef(fit)

}
