% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtn}
\alias{rtn}
\title{Daily log return of the GSPC from 2000-02-01 to 2002-06-28}
\format{
A vector with length 604.
\describe{
  \item{log return}{log return}
  ...
}
}
\source{

}
\usage{
rtn
}
\description{
This is processed from the original ^GSPC.csv dataframe.
}
\keyword{datasets}
